Make .Debug flatpak more useful
Since in the discussion of the initial pull request for the USDX flatpak
it was requested to build without --enable-debug, the .Debug flatpak so
far contained only debug info for the two small C++ API wrappers. With
this patch a release build of USDX will generate debug info if the CFLAGS
contain -g.